pyproject.toml
[tool.poetry]
name = "smart-mini-crt-interface"
version = "0.1.0"
description = ""
authors = ["Will Garside <worgarside@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
pigpio = "*"
pillow = "*"
tk = "*"
wg-utilities = { extras = ["devices-dht22", "mqtt"], version = "^5.16.0" }

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.mypy]
exclude = ["_local_sandbox"]
mypy_path = "smart_mini_crt_interface"

follow_imports = "normal"
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_decorators = true
disallow_untyped_defs = true

warn_return_any = true
warn_unused_ignores = true
warn_unused_configs = true
warn_redundant_casts = true

no_implicit_optional = true
strict_optional = true

strict_equality = true
explicit_package_bases = true
show_error_codes = true
enable_error_code = [
  "arg-type",           # Checks for argument types in function calls.
  "assignment",         # Validates types of variable assignments.
  "return-value",       # Ensures that return values match the declared type.
  "call-overload",      # Validates correct use of method/function overloads.
  "operator",           # Checks operator usage against expected types.
  "index",              # Validates types used in indexing operations.
  "attr-defined",       # Checks if attributes exist in classes/objects.
  "no-untyped-call",    # Prevents calls to untyped functions in typed contexts.
  "unreachable",        # Warns about code that can never be executed.
  "truthy-bool",        # Ensures that boolean expressions are boolean in type.
  "union-attr",         # Checks for attribute access in unions safely.
  "dict-item",          # Checks for proper types used in dictionary items.
  "unused-awaitable",   # Warns if an awaitable is not awaited.
  "override",           # Ensures overridden methods match the signature in base class.
  "list-item",          # Checks type consistency of list items.
  "return",             # Ensures all code paths in a function return a value if needed.
  "type-var",           # Ensures type variables are used correctly.
  "valid-type",         # Validates types used in annotations are recognized.
  "func-returns-value", # Ensures functions declared with a return type actually return a value.
]

[tool.codespell]
skip = '*.json,*.csv,*.xml,*.svg,*.lock'

[tool.ruff]
fix = true
preview = true
line-length = 90
lint.select = ["ALL"]
lint.ignore = [
  "ANN401", # https://docs.astral.sh/ruff/rules/any-type/
  "D107",   # https://docs.astral.sh/ruff/rules/undocumented-public-init/
  "S314",   # https://docs.astral.sh/ruff/rules/suspicious-xml-element-tree-usage/
  "TRY003", # https://docs.astral.sh/ruff/rules/raise-vanilla-args/
  "ISC001", # https://docs.astral.sh/ruff/rules/single-line-implicit-string-concatenation/
  "TD002",  # https://docs.astral.sh/ruff/rules/missing-todo-author/
  "TD003",  # https://docs.astral.sh/ruff/rules/missing-todo-link/
  "EM",     # https://docs.astral.sh/ruff/rules/#flake8-errmsg-em
  "FIX002", # https://docs.astral.sh/ruff/rules/line-contains-todo/
  "ANN101", # https://docs.astral.sh/ruff/rules/missing-type-self/
  "CPY",    # https://docs.astral.sh/ruff/rules/#flake8-copyright-cpy
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["D104"]

[tool.ruff.lint.pycodestyle]
max-line-length = 115

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.pylint]
max-args = 10

[tool.ruff.lint.isort]
required-imports = ["from __future__ import annotations"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = "dynamic"

[tool.vulture]
min_confidence = 70
paths = ["smart_mini_crt_interface"]
sort_by_size = true

[tool.creosote]
paths = ["smart_mini_crt_interface"]
deps-file = "pyproject.toml"
sections = ["tool.poetry.dependencies"]
exclude-deps = ["paho-mqtt", "pigpio", "pillow", "tk"]
